requirements.txt
# Voice-First Agentic AI System - Requirements

# Core dependencies
google-generativeai>=0.3.0

# Speech processing
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.11.0
gTTS>=2.5.0
pygame>=2.5.0
pyttsx3>=2.90  # Offline TTS fallback

# Audio utilities (optional fallback)
playsound>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.0